refactor: 예외 코드 error Tyep에서 순수 예외 클래스로 변경

+++ modules/api-dashboard/src/main/kotlin/com/goodpon/api/dashboard/api/controller/v1/request/CreateCouponTemplateWebRequest.kt
@@ -9,7 +9,7 @@ import java.time.LocalDate
 data class CreateCouponTemplateWebRequest(
     val name: String,
     val description: String,
-    val minOrderAmount: Long?,
+    val minOrderAmount: Int?,
     val discountType: CouponDiscountType,
     val discountValue: Int,
     val maxDiscountAmount: Int?,
