+++ modules/core/src/main/kotlin/com/goodpon/core/application/coupon/request/CreateCouponTemplateRequest.kt
@@ -12,7 +12,7 @@ data class CreateCouponTemplateRequest(
     val name: String,
     val description: String,
     val merchantId: Long,
-    val minOrderAmount: Long?,
+    val minOrderAmount: Int?,
     val discountType: CouponDiscountType,
     val discountValue: Int,
     val maxDiscountAmount: Int?,
