+++ modules/infra/jpa/src/main/kotlin/com/goodpon/infra/jpa/coupon/CouponTemplateEntity.kt
@@ -21,7 +21,7 @@ class CouponTemplateEntity(
     val description: String,
 
     @Column
-    val minOrderAmount: Long? = null,
+    val minOrderAmount: Int? = null,
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
